add edge loss

+++ model/model_mobilenetv2_seg_small.py
@@ -39,14 +39,14 @@ def conv_1x1_bn(inp, oup):
 
 def conv_bn(inp, oup, kernel, stride):
     return nn.Sequential(
-        nn.Conv2d(in_channels=inp, out_channels=oup, kernel_size=kernel, stride=stride, padding=(kernel-1)/2, bias=False),
+        nn.Conv2d(in_channels=inp, out_channels=oup, kernel_size=kernel, stride=stride, padding=int((kernel-1)/2), bias=False),
         nn.BatchNorm2d(num_features=oup, eps=1e-05, momentum=0.1, affine=True),
         nn.ReLU(inplace=True)
     )
 
 def conv_dw(inp, oup, kernel, stride):
     return nn.Sequential(
-        nn.Conv2d(inp, inp, kernel, stride, (kernel-1)/2, groups=inp, bias=False),
+        nn.Conv2d(inp, inp, kernel, stride, int((kernel-1)/2), groups=inp, bias=False),
         nn.BatchNorm2d(num_features=inp, eps=1e-05, momentum=0.1, affine=True),
         nn.ReLU(inplace=True),
